Add capability to add a song index
This adds a song index to the songbook.
Make now looks for index files and builds them.
using the songidx lua script provided by the songs package.
Note that the makefile expects the script to be in /usr/share/songs.
This maybe only works on Arch Linux with the songs package.
Therefore the script should be included in the repo in the future.